Video Author / Protecting Routes / Add Videos in UserDetail / Solve the deleteButton  & changePssword Buttom problem

+++ controller/videoController.js
@@ -41,20 +41,21 @@ export const postUpload = async (req, res) => {
     fileUrl: path,
     title,
     description,
+    creator: req.user.id,
   });
-  console.log(newVideo);
+  req.user.videos.push(newVideo.id);
+  req.user.save();
+
   res.redirect(routes.videoDetail(newVideo.id));
   //To Do : Upload and save video 위에서 받은 변수 이용
 };
 
 export const videoDetail = async (req, res) => {
-  console.log(req.params);
   const {
     params: { id },
   } = req;
-  const video = await Video.findById(id);
+  const video = await Video.findById(id).populate("creator");
   try {
-    console.log(video);
     res.render("videoDetail", { pageTitle: `${video.title}`, video });
   } catch (error) {
     console.log(error);
@@ -68,7 +69,13 @@ export const getEditVideo = async (req, res) => {
   } = req;
   try {
     const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: "editVideo", video });
+    console.log(req.user.id);
+    console.log(video.creator);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: "editVideo", video });
+    }
   } catch (error) {
     res.redirect(routes.home);
   }
@@ -93,9 +100,14 @@ export const deleteVideo = async (req, res) => {
   const {
     params: { id },
   } = req;
-
+  console.log("아이디!!!!" + id);
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator != req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
